organising/adding mvvm on detail view
