+++ vlogsApp/vlogsApp/ViewController/Detail/DetailScreenView.swift
@@ -1,0 +1,84 @@
+//
+//  DetailScreenView.swift
+//  vlogsApp
+//
+//  Created by Petar Popovski on 15.8.23.
+//
+
+import UIKit
+import SnapKit
+
+class DetailScreenView: UIView {
+    
+    var blog: Blog?
+    
+    lazy var titleLabel: UILabel = {
+        let label = UILabel()
+        label.textColor = UIColor(named: "titleColor")
+        label.font = UIFont.boldSystemFont(ofSize: 18)
+        label.textAlignment = .center
+        label.numberOfLines = 0
+        return label
+    }()
+    
+    lazy var descriptionLabel: UILabel = {
+        let label = UILabel()
+        label.textColor = UIColor(named: "titleColor")
+        label.font = UIFont.systemFont(ofSize: 14)
+        label.textAlignment = .center
+        label.numberOfLines = 0
+        return label
+    }()
+    
+    lazy var vlogImageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.clipsToBounds = true
+        imageView.layer.cornerRadius = 20
+        imageView.layer.borderWidth = 1
+        imageView.contentMode = .scaleToFill
+        return imageView
+    }()
+    
+    lazy var vlogImage: UIImage = {
+        let image = UIImage()
+        return image
+    }()
+    
+    override init (frame: CGRect) {
+        super.init(frame: .zero)
+        
+        configUI()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    func configUI() {
+        
+        backgroundColor = UIColor(named: "backgroundColor")
+        
+        addSubview(titleLabel)
+        addSubview(descriptionLabel)
+        addSubview(vlogImageView)
+        
+        titleLabel.snp.makeConstraints { make in
+            make.centerX.equalToSuperview()
+            make.top.equalTo(vlogImageView.snp.bottom).offset(20)
+            make.leading.trailing.equalToSuperview().inset(12)
+        }
+        
+        descriptionLabel.snp.makeConstraints { make in
+            make.centerX.equalToSuperview()
+            make.top.equalTo(titleLabel.snp.bottom).offset(10)
+            make.leading.trailing.equalToSuperview().inset(12)
+        }
+        
+        vlogImageView.snp.makeConstraints { make in
+            make.centerX.equalToSuperview()
+            make.top.equalToSuperview().offset(20)
+            make.width.equalTo(350)
+            make.height.equalTo(250)
+        }
+    }
+}
